requirements.txt
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==1.4.0
attrs==20.2.0
Babel==2.8.0
bcrypt==3.2.0
black==20.8b1
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cliff==3.1.0
cmd2==0.8.9
colorama==0.4.3
cryptography==3.1
debtcollector==1.22.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.1
django-bootstrap3==12.1.0
django-extensions==3.0.9
django-phonenumber-field==1.1.0
django-pipeline==2.0.5
django-slack==5.15.2
dogpile.cache==1.0.0
ecdsa==0.16.0
fabric==2.5.0
filelock==3.0.12
flake8==3.8.3
funcsigs==1.0.2
gunicorn==20.0.4
identify==1.5.5
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
invoke==1.4.1
ipaddress==1.0.23
iso8601==0.1.13
jmespath==0.10.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
keystoneauth1==4.1.0
mccabe==0.6.1
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
munch==2.5.0
mypy-extensions==0.4.3
netaddr==0.8.0
netifaces==0.10.9
nodeenv==1.5.0
openstacksdk==0.47.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.0.0
oslo.config==7.0.0
oslo.context==2.23.1
oslo.i18n==3.25.1
oslo.log==3.45.2
oslo.serialization==2.29.2
oslo.utils==3.42.1
paramiko==2.7.2
pathspec==0.8.0
pbr==5.5.0
phonenumberslite==8.12.9
positional==1.2.1
pre-commit==2.7.1
prettytable==0.7.2
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pyflakes==2.2.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.17.3
python-cinderclient==6.0.0
python-dateutil==2.8.1
python-glanceclient==3.0.0
python-keystoneclient==3.22.0
python-neutronclient==7.1.0
python-novaclient==16.0.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requestsexceptions==1.4.0
rfc3986==1.4.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
sshpubkeys==3.1.0
sshuttle==1.0.4
stevedore==3.2.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.10
virtualenv==20.0.31
warlock==1.3.3
wcwidth==0.2.5
wrapt==1.12.1
zipp==1.2.0
